Migrate ``tests/unit/modules/test_dockermod.py`` to Pytest
